# Conflicts:
#	core/src/main/java/org/frankframework/jms/MessagingSource.java
#	core/src/main/java/org/frankframework/jta/btm/BtmDataSourceFactory.java

+++ core/src/main/java/org/frankframework/filesystem/Samba1FileSystem.java
@@ -156,13 +156,14 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 		}
 	}
 
-	public boolean isFolder(SmbFile f) throws FileSystemException {
+	private boolean isFolder(SmbFile f) throws FileSystemException {
 		try {
 			return f.isDirectory();
 		} catch (SmbException e) {
 			throw new FileSystemException(e);
 		}
 	}
+
 	@Override
 	public boolean folderExists(String folder) throws FileSystemException {
 		return isFolder(toFile(folder));
@@ -189,9 +190,13 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 		String normalized = FilenameUtils.normalizeNoEndSeparator(folder, true) + "/";
 		try {
 			if (folderExists(normalized)) {
+				if(!removeNonEmptyFolder && listFiles(folder).iterator().hasNext()) {
+					throw new FileSystemException("Cannot remove folder [" + folder + "]. Directory not empty.");
+				}
+
 				toFile(normalized).delete();
 			} else {
-				throw new FileSystemException("Remove directory for [" + normalized + "] has failed. Directory does not exist.");
+				throw new FileSystemException("Cannot remove folder [" + normalized + "]. Directory does not exist.");
 			}
 		} catch (SmbException e) {
 			throw new FileSystemException(e);
